refactor: add new folders and improve imports
